Update robotics.py
change the return type of DriveBase.angle() to be a float in the type stub

+++ src/pybricks/robotics.py
@@ -101,8 +101,8 @@ class DriveBase:
             Driven distance since last reset.
         """
 
-    def angle(self) -> int:
-        """angle() -> int: deg
+    def angle(self) -> float:
+        """angle() -> float: deg
 
         Gets the estimated rotation angle of the drive base.
